move the pkg
